Introducing a signal handler watcher thread
Some applications install their own signal handlers. If this takes place
after GFXR installs the signal handler that is part of the page_guard
memory tracking method, then it effectively renders the mechanism
useless and cause segmantation faults.

This patch introduces a which constantly monitors the active SIGSEG
handler, and it is not the one installed by the page guard mechanism
then replaces it back.

This feature is disabled by default and can be enabled by setting the
appropriate enviroment variable, depending on the OS.