screenshot: Add support to export PNG format
Add the ability for grabbing PNG formatted screenshots during replay and
D3D12 capture.  The PNG format is using STB headers pulled in as a
submodule with the additional tweak of using the Zlib we already pull
in to increase the compression ratio versus the original image.

The current implementation supports 2 modes: threaded, non-threaded.
For desktop replay, the replay will be multithreaded to 4 threads
(defined in the CMakelists.txt file using the MAX_SCREENSHOT_THREADS
define which is passed into the compiler definitions).
For Android replay, the replay will be single-threaded to not put
any additional memory pressure on the Android device.