Fixes in VulkanResourcesUtil
Fixes in the scale image path:
- Pipeline barrier after Blit was using stale/wrong srcAccessMask and
dstAccessMask
- Some artifacts are seen on amd 7800 and on the S22 when Blit command
uses LINEAR filter. Switching to NEAREST appears to be fixing these
artifacts
- Stricter checks are performed if the implementation supports the
requested combination of image format and tiling. It is possible that an
implementation does not support to perform blit in some cases.
- Since scale is allowed to be greater than 1.0f the scaled image
dimensions need to be checked with the limits impossed by the
implementation

Other fixes in VulkanResourcesUtil:
- Fixes in some src access masks and stages