Fix parsing of txt dump resources args when some params are missing
If a draw index is not specified in the first line of the txt dump resources
arg file (because the dump is for a tracerays or dispatch call), and is used
in a subsequent line, the vector holding the dump indices was not initialized
correctly, resulting in a crash. This change fixes the problem by always
initializing each of dump indices vector elements to empty. This was also
done for the dispatch and tracerays vectors.

Also fixed wording in some help text/documentation.