VDR: Move some context specific code from common
VulkanReplayDumpResourcesBase class serves as a filtering layer between
the replay consumer and the dump resources contexts (draw
calls/compute/ray tracing) that handle the dumping. Its job is to filter
the api calls and forward it into the appropriate dumping contexts.

This commit moves some code that is too specific for the
VulkanReplayDumpResourcesBase class and places it into the appropriate
dumping context classes.