VDR: Simplify IsRecodring
IsRecording()'s job is to guard the resource dumping logic to be called
when not actually dumping anything. In its current logic it searches
among available dumping contexts in an inefficient and complicated way
which is not 100% correct.

This commit simplifies the logic by keeping track of the currently
active dumping contexts.